Fix #4: Tailor's original data are correct
Here are excerpts from Tailor's email explaining:

The centrifuge temps that were missing on the data sheets came from me noticing the
centrifuge stayed 18 the whole time I used it.

The lengths I entered for 1016 and 1017 were incorrect on the data sheet that I
gave you to enter and correct on the data that I entered on Github. I had
measured the lengths and wrote them down on a piece of paper next to my stem
cutting tub since the sheets do not have a stem length column. I had been
writing them down on the data sheet after I finish the stem. (which is a dumb
method) So I was writing down stem lengths for another stem, but didn't
correctly document which stem. It was for stem P3307 which is a stem that needs
to be entered into the data on Github. I have 1016 in the refrigerator and
double checked its length with what I had entered on github.