Data sent by @marleyclifton on 2014-09-08
- Data added each file
- protein-leaves.csv file line endings fixed

Here is the text of the email Marley sent on 2014-09-08 along with these
data:

"This is the updated data for all of the trees except the last trip to
the CM. A few things to note, there are a few places where I have a tree
tag but no number. For Jude 157 I am fairly positive this is a tree we
tagged in the DM but did not collect. I remember this one specifically
because I was in the field with you and we didn’t have time to go back.
There is only one oak sample i did not have protein for in my notes. But
I went through every bag and did protein before I did leaf mass/ area so
I will have to manually look in the bag so I can enter the data on that
sample. There are three pines I did not have in my notes on my computer.
I am not home so I will go back and look at my written notes for those
pines dimensions, I know I have them because I have weights for the
measured pines."