Update pine dimensions from @marleyclifton
Data emailed by @marleyclifton on 2014-09-09.  From her email:

"I had pist 1037 labeled as 1027, I had two 1027s but one was a pied and
the other was a pist3. I renamed the tag correctly and put it in the
correct order.

The alignment of the needles column got off when I added two blank tree
lines, that is now fixed.

I added the pines dimensions for PICE 1151 and 1159."